requirements.txt
Django==3.1.14
django-constance==2.7.0
django-cors-headers==3.5.0
django-chunked-upload==2.0.0
django-cryptography==1.0
django-extensions==3.0.9
django-filter==2.4.0
django-picklefield==3.0.1
django-redis==4.12.1
django-rq==2.5.1
django-bulk-update
django-silk==4.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
drf-extensions==0.7.0
drf-spectacular==0.21.2
face-recognition==1.2.3
faiss-cpu==1.7.2
Flask==1.1.4
Flask-Cors==3.0.9
Flask-RESTful==0.3.9
geopy==2.1.0
gunicorn==20.1.0
hdbscan==0.8.28
matplotlib==3.3.2 
networkx==2.6.3
nltk==3.6.7
markupsafe==2.0.1
Pillow==9.0.1
psycopg2==2.8.6
pycocotools==2.0.2
https://github.com/owncloud/pyocclient/archive/master.zip
pytz==2020.1
PyExifTool==0.4.9
PyJWT==2.4.0
pyvips==2.2.1
rq==1.10.1
seaborn==0.11.0
sentence_transformers==1.2.0
sklearn==0.0
timezonefinder==5.2.0
tqdm==4.51.0
gevent==21.12.0
python-magic==0.4.18
Wand==0.6.6 
